requirements.txt
# YouTube video downloading
yt-dlp>=2023.10.13

# AWS S3 integration
boto3>=1.28.0
botocore>=1.31.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests for API calls
requests>=2.31.0

# REST API server
flask>=3.0.0
flask-cors>=4.0.0

# Optional: For better progress bars (yt-dlp has built-in support)
# tqdm>=4.65.0

